Make one instance of database.
